- fixing various MSVC portability issues - fixing a few assertions which were reported by our sanity checking assertion code: each assert should execute a check that produces a pure boolean result.

+++ cxxmph/bfcr_map_test.cc
@@ -65,14 +65,14 @@ int main(int argc, char** argv) {
        if (key < num_valid && it->second != key) exit(-1);
     }
   }
-  fprintf(stderr, "Size is %lu\n", h.size());
+  fprintf(stderr, "Size is %lu\n", (unsigned long)h.size());
   for (int i = 0; i < num_valid; ++i) {
      char buf[10];    
      snprintf(buf, 10, "%ld", i);
      h.erase(buf);
   }
   if (h.size() != 1) {
-    fprintf(stderr, "Erase failed: size is %lu\n", h.size());
+    fprintf(stderr, "Erase failed: size is %lu\n", (unsigned long)h.size());
     exit(-1);
   }
 }
